UI Kit remove mixins (#2297)
* Remove mixin from ui kit text component

* Remove mixins from grid-item component

* Move spacing and border mixins into box component

* Fix displayclass object object issue. Fix link spacing issue

* Add test for an edge case I found

* Fix visual issues. Fix edge case for responsive prop util

* Simplify tvj test
