+++ app/components/ui-kit/box.js
@@ -1,22 +1,36 @@
 import Component from '@ember/component';
 import { reads } from '@ember/object/computed';
 
-import spacingMixin from 'travis/mixins/ui-kit/spacing';
-import borderMixin from 'travis/mixins/ui-kit/border';
 import { checkDictionary } from 'travis/utils/ui-kit/assertions';
 import prefix from 'travis/utils/ui-kit/prefix';
 import concat from 'travis/utils/ui-kit/concat';
 
 const COLORS = {
   WHITE: 'white',
+  GREY_LIGHT: 'grey-light',
 };
 
 const BG_COLORS = {
   [COLORS.WHITE]: 'white',
 };
 
-const DISPLAYS = {
+const BORDER_COLORS = {
+  [COLORS.GREY_LIGHT]: 'grey-150',
+};
+
+const WIDTHS = {
+  NONE: 'none',
+  XS: 'xs',
+};
+
+const BORDER_WIDTHS = {
+  [WIDTHS.NONE]: 'none',
+  [WIDTHS.XS]: 'px',
+};
+
+export const DISPLAYS = {
   BLOCK: 'block',
+  INLINE: 'inline',
   INLINE_BLOCK: 'inline-block',
   FLEX: 'flex',
 };
@@ -93,7 +107,7 @@ const MAX_WIDTHS = {
 
 
 // Component definition
-export default Component.extend(spacingMixin, borderMixin, {
+export default Component.extend({
   tagName: '',
 
   // Public interface //
@@ -120,7 +134,7 @@ export default Component.extend(spacingMixin, borderMixin, {
 
   // Private //
   colorClass: prefix('color', 'bg', { dictionary: BG_COLORS }),
-  displayClass: reads('display'),
+  displayClass: prefix('display', ''),
   layerClass: prefix('layer', 'z'),
   overflowAllClass: prefix('overflow.all', 'overflow'),
   overflowXClass: prefix('overflow.x', 'overflow'),
@@ -140,6 +154,57 @@ export default Component.extend(spacingMixin, borderMixin, {
   positionBottom: prefix('position.bottom', 'bottom'),
   positionLeft: prefix('position.left', 'left'),
   positionInset: prefix('position.inset', 'inset'),
+
+  // Border
+  borderAllWidthClass: prefix('borderWidth.all', 'border', { dictionary: BORDER_WIDTHS }),
+  borderTopWidthClass: prefix('borderWidth.top', 'border-t', { dictionary: BORDER_WIDTHS }),
+  borderRightWidthClass: prefix('borderWidth.right', 'border-r', { dictionary: BORDER_WIDTHS }),
+  borderBottomWidthClass: prefix('borderWidth.bottom', 'border-b', { dictionary: BORDER_WIDTHS }),
+  borderLeftWidthClass: prefix('borderWidth.left', 'border-l', { dictionary: BORDER_WIDTHS }),
+  borderColorClass: prefix('borderColor', 'border', { dictionary: BORDER_COLORS }),
+  borderWidthClasses: concat(
+    'borderAllWidthClass',
+    'borderTopWidthClass',
+    'borderRightWidthClass',
+    'borderBottomWidthClass',
+    'borderLeftWidthClass',
+  ),
+
+  // Margin
+  marginTop: prefix('margin.top', 'mt', { negatable: true }),
+  marginRight: prefix('margin.right', 'mr', { negatable: true }),
+  marginBottom: prefix('margin.bottom', 'mb', { negatable: true }),
+  marginLeft: prefix('margin.left', 'ml', { negatable: true }),
+  marginX: prefix('margin.x', 'mx', { negatable: true }),
+  marginY: prefix('margin.y', 'my', { negatable: true }),
+  marginAll: prefix('margin.all', 'm', { negatable: true }),
+  marginClasses: concat(
+    'marginTop',
+    'marginRight',
+    'marginBottom',
+    'marginLeft',
+    'marginX',
+    'marginY',
+    'marginAll',
+  ),
+
+  // Padding
+  paddingTop: prefix('padding.top', 'pt'),
+  paddingRight: prefix('padding.right', 'pr'),
+  paddingBottom: prefix('padding.bottom', 'pb'),
+  paddingLeft: prefix('padding.left', 'pl'),
+  paddingX: prefix('padding.x', 'px'),
+  paddingY: prefix('padding.y', 'py'),
+  paddingAll: prefix('padding.all', 'p'),
+  paddingClasses: concat(
+    'paddingTop',
+    'paddingRight',
+    'paddingBottom',
+    'paddingLeft',
+    'paddingX',
+    'paddingY',
+    'paddingAll',
+  ),
 
   // Collected classes
   allClasses: concat(
@@ -191,5 +256,14 @@ export default Component.extend(spacingMixin, borderMixin, {
     checkDictionary(all, OVERFLOWS, '@overflow.all', 'Box');
     checkDictionary(x, OVERFLOWS, '@overflow.x', 'Box');
     checkDictionary(y, OVERFLOWS, '@overflow.y', 'Box');
+
+    checkDictionary(this.borderColor, COLORS, '@borderColor');
+
+    const bw = this.borderWidth || {};
+    checkDictionary(bw.top, WIDTHS, '@borderWidth.top');
+    checkDictionary(bw.right, WIDTHS, '@borderWidth.right');
+    checkDictionary(bw.bottom, WIDTHS, '@borderWidth.bottom');
+    checkDictionary(bw.left, WIDTHS, '@borderWidth.left');
+    checkDictionary(bw.all, WIDTHS, '@borderWidth.all');
   },
 });
