+++ app/components/ui-kit/grid-item.js
@@ -2,8 +2,6 @@ import Component from '@ember/component';
 import { computed } from '@ember/object';
 import { or } from '@ember/object/computed';
 
-import spacingMixin from 'travis/mixins/ui-kit/spacing';
-import borderMixin from 'travis/mixins/ui-kit/border';
 import { requireProp } from 'travis/utils/ui-kit/assertions';
 import concat from 'travis/utils/ui-kit/concat';
 
@@ -38,7 +36,7 @@ const FLEX_SIZE_VALS = {
   [FLEX_SIZES.NONE]: 'flex-none',
 };
 
-export default Component.extend(spacingMixin, borderMixin, {
+export default Component.extend({
   tagName: '',
 
   // Public interface
@@ -51,6 +49,12 @@ export default Component.extend(spacingMixin, borderMixin, {
   lg: null,
   xl: null,
   gap: null,
+
+  display: null,
+  borderColor: null,
+  borderWidth: null,
+  margin: null,
+  padding: null,
 
   // Private
   sizePrefix: computed('grid.isCol', function () {
@@ -84,10 +88,6 @@ export default Component.extend(spacingMixin, borderMixin, {
     'lgClass',
     'xlClass',
     'gapClass',
-    'borderColorClass',
-    'borderWidthClasses',
-    'marginClasses',
-    'paddingClasses',
   ),
 
   // Lifecycle
