+++ app/templates/components/ui-kit/box.hbs
@@ -1,6 +1,6 @@
-{{#let (element this.tag) as |Tag|}}
+{{#let (element this.tag) as |Tag|~}}
   <Tag
     class={{this.allClasses}}
     ...attributes
   >{{yield}}</Tag>
-{{/let}}
+{{~/let}}
