+++ app/templates/components/ui-kit/link.hbs
@@ -1,10 +1,10 @@
-<a
-  class={{this.allClasses}}
+<UiKit::Text
+  @tag='a'
+  @color={{this.color}}
   href={{this.href}}
   rel={{this.rel}}
   target={{this.target}}
   ...attributes
-><UiKit::Text
-    @tag='span'
-    @color={{this.color}}
-  >{{yield}}</UiKit::Text></a>
+>
+  {{~yield~}}
+</UiKit::Text>
