+++ app/templates/components/ui-kit/text.hbs
@@ -1,8 +1,10 @@
-{{#let (element this.tag) as |Tag|~}}
-  <Tag
+<UiKit::Box
+  @tag={{this.tag}}
+  @display={{this.display}}
+  @margin={{this.margin}}
+  @padding={{this.padding}}
   class={{this.allClasses}}
   ...attributes
-  >
-    {{~yield~}}
-  </Tag>
-{{~/let}}
+>
+  {{~yield~}}
+</UiKit::Box>
